# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RestReviews API - Restoranlar va Sharhlar",
        "description": "Mobile ilova uchun REST API. Restoranlarni kashf qilish, sharh qoldirish, sevimlilar va boshqalar.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gastronomic.webclub.uz",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://restreviews.test",
            "description": "Laragon Server"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "👤 Profil"
                ],
                "summary": "Client profil ma'lumotlari",
                "description": "Login qilgan clientning to'liq profil ma'lumotlari va statistikasi",
                "operationId": "f1784685f8270527da017b3616288443",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en). Default: kk",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "kk",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil ma'lumotlari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Abdulla"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Valiyev"
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Abdulla Valiyev"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "998901234567"
                                                },
                                                "image_path": {
                                                    "type": "string",
                                                    "example": "https://example.com/storage/clients/photo.jpg",
                                                    "nullable": true
                                                },
                                                "statistics": {
                                                    "properties": {
                                                        "total_reviews": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "five_star_reviews": {
                                                            "type": "integer",
                                                            "example": 8
                                                        },
                                                        "favorites_count": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "reviews": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "👤 Profil"
                ],
                "summary": "Profil ma'lumotlarini yangilash",
                "description": "Client ism, familiya va rasmini yangilash",
                "operationId": "df42d7bd8c9605adf2e62002661ca9f3",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "Ism",
                                        "type": "string",
                                        "example": "Abdulla",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "description": "Familiya",
                                        "type": "string",
                                        "example": "Valiyev",
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Profil rasmi (maksimum 2MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil muvaffaqiyatli yangilandi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil muvaffaqiyatli yangilandi"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}